\name{TCGAdownload}
\alias{TCGAdownload}
\title{download TCGA data}
\usage{
TCGAdownload("TCGA-STAD", "Transcriptome Profiling","Gene Expression Quantification","HTSeq - FPKM")
}
\description{
"HTSeq - Counts"
}
\examples{
TCGAdownload("TCGA-STAD", "Transcriptome Profiling","Gene Expression Quantification","HTSeq - FPKM")}


